Clean up GA and Ext
