+++ Project.toml
@@ -28,7 +28,7 @@ ConstraintDomains = "0.3"
 Dictionaries = "0.4"
 Distances = "0.10"
 ExproniconLite = "0.10.13"
-JuliaFormatter = "1"
+JuliaFormatter = "1, 2"
 OrderedCollections = "1"
 Random = "1"
 TestItems = "1"
